Ensure we track inputs
